+++ mlflow/runs.py
@@ -7,7 +7,7 @@ import click
 import json
 import mlflow.tracking
 from mlflow.entities import ViewType
-from mlflow.tracking import _get_store
+from mlflow.store import get_tracking_store
 from tabulate import tabulate
 from mlflow.utils.mlflow_tags import MLFLOW_RUN_NAME
 from mlflow.utils.time_utils import conv_longdate_to_str
@@ -34,7 +34,7 @@ def list_run(experiment_id, view):
     """
     List all runs of the specified experiment in the configured tracking server.
     """
-    store = _get_store()
+    store = get_tracking_store()
     view_type = ViewType.from_string(view) if view else ViewType.ACTIVE_ONLY
     runs = store.search_runs([experiment_id], None, view_type)
     table = []
@@ -53,7 +53,7 @@ def delete_run(run_id):
     is already marked. You can restore a marked run with ``restore_run``,
     or permanently delete a run in the backend store.
     """
-    store = _get_store()
+    store = get_tracking_store()
     store.delete_run(run_id)
     print("Run with ID %s has been deleted." % str(run_id))
 
@@ -65,7 +65,7 @@ def restore_run(run_id):
     Restore a deleted run.
     Returns an error if the run is active or has been permanently deleted.
     """
-    store = _get_store()
+    store = get_tracking_store()
     store.restore_run(run_id)
     print("Run with id %s has been restored." % str(run_id))
 
@@ -76,7 +76,7 @@ def describe_run(run_id):
     """
     All of run details will print to the stdout as JSON format.
     """
-    store = _get_store()
+    store = get_tracking_store()
     run = store.get_run(run_id)
     json_run = json.dumps(run.to_dictionary(), indent=4)
     print(json_run)
