+++ mlflow/server/handlers.py
@@ -23,8 +23,8 @@ from mlflow.protos.model_registry_pb2 import ModelRegistryService, CreateRegiste
 from mlflow.protos.databricks_pb2 import RESOURCE_DOES_NOT_EXIST
 from mlflow.store.artifact.artifact_repository_registry import get_artifact_repository
 from mlflow.store.db.db_types import DATABASE_ENGINES
-from mlflow.tracking._model_registry.registry import ModelRegistryStoreRegistry
-from mlflow.tracking._tracking_service.registry import TrackingStoreRegistry
+from mlflow.store.model_registry.registry import ModelRegistryStoreRegistry
+from mlflow.store.tracking.registry import TrackingStoreRegistry
 from mlflow.utils.proto_json_utils import message_to_json, parse_dict
 from mlflow.utils.validation import _validate_batch_log_api_req
 from mlflow.utils.string_utils import is_string_type
