+++ mlflow/store/__init__.py
@@ -1,10 +1,17 @@
 from mlflow.store.tracking import abstract_store
+from mlflow.store.tracking.registry import get_tracking_store, register_tracking_store
+from mlflow.store.model_registry.registry import (
+    get_model_registry_store, register_model_registry_store
+)
 from mlflow.store.artifact import artifact_repo
 
 __all__ = [
     # tracking server meta-data stores
     "abstract_store",
-
     # artifact repository stores
     "artifact_repo",
+    "get_tracking_store",
+    "register_tracking_store",
+    "get_model_registry_store",
+    "register_model_registry_store",
 ]
