+++ mlflow/store/artifact/dbfs_artifact_repo.py
@@ -4,10 +4,10 @@ import json
 
 from mlflow.entities import FileInfo
 from mlflow.exceptions import MlflowException
+from mlflow.store import get_tracking_store
 from mlflow.store.tracking.rest_store import RestStore
 from mlflow.store.artifact.artifact_repo import ArtifactRepository
 from mlflow.store.artifact.local_artifact_repo import LocalArtifactRepository
-from mlflow.tracking._tracking_service import utils
 from mlflow.utils.file_utils import relative_path_to_artifact_path
 from mlflow.utils.rest_utils import http_request, http_request_safe, RESOURCE_DOES_NOT_EXIST
 from mlflow.utils.string_utils import strip_prefix
@@ -135,7 +135,7 @@ class DbfsRestArtifactRepository(ArtifactRepository):
 
 
 def _get_host_creds_from_default_store():
-    store = utils._get_store()
+    store = get_tracking_store()
     if not isinstance(store, RestStore):
         raise MlflowException('Failed to get credentials for DBFS; they are read from the ' +
                               'Databricks CLI credentials or MLFLOW_TRACKING* environment ' +
