+++ mlflow/tracking/__init__.py
@@ -10,7 +10,6 @@ from mlflow.tracking._tracking_service.utils import (
     set_tracking_uri,
     get_tracking_uri,
     is_tracking_uri_set,
-    _get_store,
     _TRACKING_URI_ENV_VAR
 )
 from mlflow.tracking.fluent import _EXPERIMENT_ID_ENV_VAR, _EXPERIMENT_NAME_ENV_VAR, \
@@ -21,7 +20,6 @@ __all__ = [
     "get_tracking_uri",
     "set_tracking_uri",
     "is_tracking_uri_set",
-    "_get_store",
     "_EXPERIMENT_ID_ENV_VAR",
     "_RUN_ID_ENV_VAR",
     "_TRACKING_URI_ENV_VAR",
